refactor(integrands): move find_inner/boundary_intersection helpers to base class
Deduplicate the identical find_inner_intersection() and
find_boundary_intersection() helpers that appeared in all three
intersection integrand test files by hoisting them into IntegrandTest<G>
in integrands.hh. Also adds the required rangegenerators include there.

Fixes SonarCloud duplication quality gate on PR #293.
